feat: new domain dialog
