+++ src/app/src/DomainsPage/NewDomainDialog.vue
@@ -1,0 +1,45 @@
+<script setup lang="ts">
+import InputText from 'primevue/inputtext'
+import Select from 'primevue/select'
+import Button from 'primevue/button'
+import { ref } from 'vue'
+import Dialog from 'primevue/dialog'
+
+defineProps({
+  isVisible: {
+    type: Boolean,
+    required: true
+  }
+})
+const selectedCity = ref()
+const cities = ref([
+  { name: 'New York', code: 'NY' },
+  { name: 'Rome', code: 'RM' },
+  { name: 'London', code: 'LDN' },
+  { name: 'Istanbul', code: 'IST' },
+  { name: 'Paris', code: 'PRS' }
+])
+</script>
+
+<template>
+  <Dialog v-bind:visible="isVisible" modal header="New Domain" :style="{ width: '30rem' }">
+    <div>
+      <form class="flex flex-col justify-center items-center gap-6">
+        <InputText placeholder="Domain" class="w-full" />
+        <Select
+          v-model="selectedCity"
+          :options="cities"
+          optionLabel="name"
+          placeholder="Select a organization"
+          class="w-full"
+        />
+        <Button
+          label="Submit"
+          severity="success"
+          class="w-full"
+          style="background-color: #00b272"
+        />
+      </form>
+    </div>
+  </Dialog>
+</template>
