+++ src/app/src/OrganizationPage/NewOrganizationDialog.vue
@@ -1,10 +1,6 @@
 <script setup lang="ts">
 import InputText from 'primevue/inputtext'
-import Select from 'primevue/select'
-import Textarea from 'primevue/textarea'
 import Button from 'primevue/button'
-import FloatLabel from 'primevue/floatlabel'
-import { ref } from 'vue'
 import Dialog from 'primevue/dialog'
 
 defineProps({
@@ -13,14 +9,6 @@ defineProps({
     required: true
   }
 })
-const selectedCity = ref()
-const cities = ref([
-  { name: 'New York', code: 'NY' },
-  { name: 'Rome', code: 'RM' },
-  { name: 'London', code: 'LDN' },
-  { name: 'Istanbul', code: 'IST' },
-  { name: 'Paris', code: 'PRS' }
-])
 </script>
 
 <template>
